Network DAO proposal adding title and description (#56)
* feat: Network DAO proposal adding title and description
* feat: Adaptive modifications for new version indexer

+++ src/TomorrowDAOServer.Domain/Vote/Index/IndexerVote.cs
@@ -18,13 +18,13 @@ public class IndexerVote
 
     public string AcceptedCurrency { get; set; }
 
-    public int ApprovedCount { get; set; }
+    public long ApprovedCount { get; set; }
 
-    public int RejectionCount { get; set; }
+    public long RejectionCount { get; set; }
 
-    public int AbstentionCount { get; set; }
+    public long AbstentionCount { get; set; }
 
-    public int VotesAmount { get; set; }
+    public long VotesAmount { get; set; }
     
-    public int VoterCount { get; set; }
+    public long VoterCount { get; set; }
 }
